--- api/celery_utils.py
@@ -1,22 +1,0 @@
-from celery import Celery
-import celery
-from flask.helpers import make_response
-from flask_utils import api
-
-def make_celery(app):
-    celery = Celery(
-        app.import_name,
-        backend='rpc://',
-        broker='pyamqp://'
-    )
-    celery.conf.update(app.config)
-
-    class ContextTask(celery.Task):
-        def __call__(self, *args, **kwargs):
-            with app.app_context():
-                return self.run(*args, **kwargs)
-
-    celery.Task = ContextTask
-    return celery
-
-celery = make_celery(api)
